move sea stuff into sea/ dir
