+++ src/databricks/sql/backend/sea/models/__init__.py
@@ -4,7 +4,7 @@ Models for the SEA (Statement Execution API) backend.
 This package contains data models for SEA API requests and responses.
 """
 
-from databricks.sql.backend.models.base import (
+from databricks.sql.backend.sea.models.base import (
     ServiceError,
     StatementStatus,
     ExternalLink,
@@ -13,7 +13,7 @@ from databricks.sql.backend.models.base import (
     ResultManifest,
 )
 
-from databricks.sql.backend.models.requests import (
+from databricks.sql.backend.sea.models.requests import (
     StatementParameter,
     ExecuteStatementRequest,
     GetStatementRequest,
@@ -23,7 +23,7 @@ from databricks.sql.backend.models.requests import (
     DeleteSessionRequest,
 )
 
-from databricks.sql.backend.models.responses import (
+from databricks.sql.backend.sea.models.responses import (
     ExecuteStatementResponse,
     GetStatementResponse,
     CreateSessionResponse,
