+++ src/databricks/sql/backend/sea/models/base.py
